% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binFreqTable.R
\name{binFreqTable}
\alias{binFreqTable}
\title{binFreqTable}
\usage{
binFreqTable(x, step)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{step}{Step size among the range of value.}
}
\value{
The bin frequency table
}
\description{
The bin frequency table is the table that represent the
frequency for a range of values in a particular variable.
If we want to create a bin frequency table then table function with cut
and breaks function can be used.
On the other hand, it may not flexible when creating a grouped bin frequency table.
}
\examples{
path <- system.file("extdata", "chr.tsv", package = "aggregatesR" )
mydf <- read.delim(path, header = TRUE, sep = "\t")
freqtable <- tapply(mydf$value, mydf$group, function(x) binFreqTable(x, step = 1000000))


}
