% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{D.stat}
\alias{D.stat}
\title{D.stat}
\usage{
D.stat(p1, p2, p3)
}
\value{
a D statistic (D varies from -1 to 1). D > 0 indicative of excess shared ancestry between P2 and P3.
}
\description{
The function D.stat is for computing the ABBA and BABA proportions at each site of a 'freq.py' table and use these to compute the D atstistic.

Further details are available on: \url{https://github.com/simonhmartin/tutorials/blob/master/ABBA_BABA_whole_genome/README.md}

freq.py: \url{https://github.com/simonhmartin/genomics_general}
}
\examples{
P1 <- "LR"
P2 <- "KQ"
P3 <- "MD"
D <- D.stat(freq[,P1], freq[,P2], freq[,P3])
print(paste("D =", round(D,4)))

}
