% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_maximum.R
\name{density_maximum}
\alias{density_maximum}
\title{inv_normalise1}
\usage{
density_maximum(data, v)
}
\arguments{
\item{x}{the variable}
}
\value{
The peak of distribution plot
}
\description{
A density plot takes a numeric variable to represent a smooth distribution curve over time. The peak of the density plot shows the maximum concentration of numeric data.
}
