% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{f.stat}
\alias{f.stat}
\title{f.stat}
\usage{
f.stat(p1, p2, p3a, p3b)
}
\description{
The D statistic provides a powerful test for introgression, but it does not quantify the proportion of the genome that has been shared. A related method has been developed to estimate f, the 'admixture proportion'.

The idea behind this approach is that we compare the observed excess of ABBA over BABA sites, to that which would be expected under complete admixture. To approximate the expectation under complete admixture we re-count ABBA and BABA but substituting a second population of the P3 species in the place of P2. If you lack a second population, you can simply split your P3 samples into two.
}
\examples{
P1 <- "LR"
P2 <- "KQ"
P3a <- "MD1"
P3b <- "MD2"
f <- f.stat(freq_table[,P1], freq_table[,P2], freq_table[,P3a], freq_table[,P3b])

}
