% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{get_block_indices}
\alias{get_block_indices}
\title{get_block_indices}
\usage{
get_block_indices(block_size, positions, chromosomes = NULL)
}
\arguments{
\item{block_size}{the block size along with chromosome}

\item{positions}{the position for each site to be analyzed.}

\item{chromosomes}{the chromosomes.}
}
\value{
a jackknife block indices
}
\description{
The function get_block_indices in the jackknife script will  the blocks, and return the 'indices' (i.e. the rows in our frequencies table) corresponding to each block.
}
\examples{
block_indices <- get_block_indices(block_size=1e6,
positions = freq_table$position,
chromosomes = freq_table$scaffold)
n_blocks <- length(block_indices)
print(paste("Genome divided into", n_blocks, "blocks."))

}
