% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{get_jackknife_sd}
\alias{get_jackknife_sd}
\title{get_jackknife_sd}
\usage{
get_jackknife_sd(block_indices, FUN, ...)
}
\arguments{
\item{block_indices}{the blocks defined by get_block_indices.}

\item{FUN}{the D statistic function (D.stat) we created earlier}
}
\value{
the standad deviation of the D statistic.
}
\description{
this function runs the jackknife procedure by calculating pseudovalues by removing one block at a time
}
\examples{
D_sd <- get_jackknife_sd(block_indices = block_indices,
FUN=D.stat,
freq_table[,P1], freq_table[,P2], freq_table[,P3])

}
