% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifFromMEME.R
\name{getMotifFromMEME}
\alias{getMotifFromMEME}
\title{Extract Motif Information from MEME Software.}
\usage{
getMotifFromMEME(data, format = "txt")
}
\arguments{
\item{data}{A txt or xml file from MEME software.}

\item{format}{txt or xml.}
}
\description{
\code{getMotifFromMEME} Extract motif information from the MEME software results.
}
\examples{

filepath <- system.file("examples", "meme.txt", package = "aggregatesR")
file <- getMotifFromMEME(data = filepath, format = "txt")
motif_seq <- getMotifFromMEME(file, format = "txt")

filepath <- system.file("examples", "meme.xml", package = "aggregatesR")
file <- getMotifFromMEME(data = filepath, format = "xml")
gene_info <- getMotifFromMEME(file, format = "xml")


}
