% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse_normal_transformation.R
\name{inv_normalise1}
\alias{inv_normalise1}
\title{inv_normalise1}
\usage{
inv_normalise1(x)
}
\arguments{
\item{x}{the phenotypes}
}
\value{
The normalized phenotype
}
\description{
Often in large GWAS the quantitative phenotype is forced to follow a Normal distribution by quantile normalisation (also inverse-Normal transfromation).

This way the effect of phenotypic outliers is diminished while keeping the ranking of the trait values constant. This also harmonizes the trait distributions across multiple cohorts by forcing them to look similar.

There are two basic ways of INT. One is to apply INT directly to the phenotype, and the INT-transformed phenotype is then regressed on genotypes and covariates. This is so called direct method.

In the in-direct method, the phenotype is first regressed on covariates to obtain residuals. Then the residuals of the regression in ascending order.

Further details are available on the supplements of FTO genotype is associated with phenotypic variability of body mass index \url{https://www.nature.com/articles/nature11401}
}
