% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse_normal_transformation.R
\name{inv_normalise2}
\alias{inv_normalise2}
\title{inv_normalise2}
\usage{
inv_normalise2(x)
}
\arguments{
\item{x}{the phenotypes}
}
\value{
The normalizedphenotype
}
\description{
inv_normalise2 is more aggressive that is suitable for the high skewed dataset.
}
