% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{plot_cov}
\alias{plot_cov}
\title{plot_cov}
\usage{
plot_cov(
  stem,
  pop_order,
  min = -0.009,
  max = 0.009,
  cex = 1,
  usemax = T,
  wcols = ""
)
}
\arguments{
\item{stem}{the given folder of the output of Treemix with the suffixes “.llik,” “.cov.gz,” and “.modelcov.gz” generated by default using Treemix.}

\item{pop_order}{A file that contains the order of pops that you wanted to be plotted.}
}
\description{
plot the cov
}
\examples{
file_path <- system.file("treemix", package = "aggregatesR" )
setwd(file_path)
par(mfrow=c(3,2))
for(edge in 0:5){
    plot_cov(stem=paste0("treemix.3.", edge),
    pop_order="pop_order.txt")
    }
}
