% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{plot_resid}
\alias{plot_resid}
\title{plot_resid}
\usage{
plot_resid(
  stem,
  pop_order,
  min = -0.009,
  max = 0.009,
  cex = 1,
  usemax = T,
  wcols = "r"
)
}
\arguments{
\item{stem}{the given folder of the output of Treemix with the suffixes “.llik,” “.cov.gz,” and “.modelcov.gz” generated by default using Treemix.}

\item{pop_order}{A file that contains the order of pops that you wanted to be plotted.}
}
\description{
Positive residuals indicate pairs of populations where the model
underestimates the observed covariance, and thus populations
where the fit might be improved by adding additional edges.
Negative residuals indicate pairs of populations where the model
overestimates the observed covariance; these are a necessary
outcome of having positive residuals, but can also sometimes be
interpreted as populations that are forced too close together due to
unmodeled migration elsewhere in the graph.
}
\examples{
file_path <- system.file("treemix", package = "aggregatesR" )
setwd(file_path)
par(mfrow=c(3, 3))
for(edge in 0:5){
    plot_resid(stem=paste0("treemix.3.", edge),
    pop_order="pop_order.txt")
    }

}
