% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{plot_tree}
\alias{plot_tree}
\title{plot_tree}
\usage{
plot_tree(
  stem,
  o = NA,
  cex = 1,
  disp = 0.003,
  plus = 0.01,
  flip = vector(),
  arrow = 0.05,
  scale = T,
  ybar = 0.1,
  mbar = T,
  plotmig = T,
  plotnames = T,
  xmin = 0,
  lwd = 1,
  font = 1
)
}
\arguments{
\item{stem}{the given folder of the output of Treemix with the suffixes “.llik,” “.cov.gz,” and “.modelcov.gz” generated by default using Treemix.}
}
\description{
Plotted is the structure of the graph inferred by TreeMix for sub-populations, allowing
different migration events. Migration arrows are colored according to their weight. Horizontal branch lengths are proportional to the amount of genetic
drift that has occurred on the branch.
}
\examples{
file_path <- system.file("treemix", package = "aggregatesR" )
setwd(file_path)
par(mfrow=c(3,3))
for(edge in 0:10){
   plot_tree(paste0("treemix.3.", edge))
   title(paste(edge,"edges"))
   }

}
