% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemix.R
\name{plot_treemix}
\alias{plot_treemix}
\title{plot_treemix}
\usage{
plot_treemix(
  obj,
  plot.nodes = TRUE,
  plot.migration = TRUE,
  branch.colour = "grey",
  label = TRUE,
  ...
)
}
\arguments{
\item{obj}{pieces of a TreeMix run from \code{read_treemix()}}

\item{plot.nodes}{show dots for internal and leaf nodes of population tree}

\item{plot.migration}{show migration edges as curved arrows}

\item{branch.colour}{colour to use when drawing edges of population tree}
}
\value{
a \code{ggplot2} plot
}
\description{
Plotted(ggplot2) is the structure of the graph inferred by TreeMix for sub-populations, allowing
different migration events. Migration arrows are colored according to their weight. Horizontal branch lengths are proportional to the amount of genetic
drift that has occurred on the branch.
}
