% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemix.R
\name{read_treemix}
\alias{read_treemix}
\title{read_treemix}
\usage{
read_treemix(stem, ...)
}
\arguments{
\item{stem}{filename stem (\code{-o} from command-line) for TreeMix output}
}
\value{
a list with components including \code{m} (number of migration edges), \code{llik} (log-likelihood)
  and \code{tree} (population tree as \code{ape::phylo} object)
}
\description{
Read TreeMix output files.
}
\examples{
for m in {0..5};
do
    for i in {1..10};
    do
    s=$RANDOM
    echo "treemix -i treemix_in.gz -o result1/treemix.${i}.${m} -root outgroup -m ${m} -global -noss -k 500 -seed ${s}"
    done
done > cmd.list
df <- read_treemix(stem = "treemix.4.2")
plot_treemix(df) + theme_bw()
}
