% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aspera_md5.R
\name{set_aspera_md5}
\alias{set_aspera_md5}
\title{set_aspera_md5}
\usage{
set_aspera_md5(path, md5_tsv)
}
\arguments{
\item{path}{the absolute path of the output file.}

\item{md5_tsv}{the ENA md5 file.}
}
\description{
Message Digest Algorithm 5(MD5) is a cryptographic hash algorithm that can be used to
verify the integrity of files.
Here, we using set_aspera_md5 to prepare
ENA(\url{https://www.ebi.ac.uk/ena/browser/home}) md5 data.
}
\examples{
path <- 'C:/Users/Admin/Desktop/'
md5_tsv <- system.file("extdata", "filereport_read_run_PRJNA648176_md5.txt", package = "aggregatesR" )
set_aspera_md5(path, md5_tsv)
}
