% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{set_y_coords}
\alias{set_y_coords}
\title{plotting_funcs}
\usage{
set_y_coords(d)
}
\description{
This R script is provided with treemix. Further details are available on the paper:
Inference of population splits and mixtures from genome-wide allele frequency data
}
\author{
Pickrell Joseph and Pritchard Jonathan
}
